Add functional worl generation
- Add small smoothing for rivers and biome heights
- Add heightmap
- Add biomemap
- Refactor World Generation

+++ src/main/java/net/jukoz/me/world/biomes/MEBiome.java
@@ -5,14 +5,14 @@ import net.minecraft.registry.RegistryKey;
 import net.minecraft.world.biome.Biome;
 
 public class MEBiome {
-    public float height;
+    public int height;
     public RegistryKey<Biome> biome;
     public Block surfaceBlock;
     public Block underSurfaceBlock;
     public Block stoneBlock;
     public Block deepStoneBlock;
 
-    public MEBiome(float height, RegistryKey<Biome> biome, Block surfaceBlock, Block underSurfaceBlock, Block stoneBlock, Block deepStoneBlock) {
+    public MEBiome(int height, RegistryKey<Biome> biome, Block surfaceBlock, Block underSurfaceBlock, Block stoneBlock, Block deepStoneBlock) {
         this.height = height;
         this.biome = biome;
         this.surfaceBlock = surfaceBlock;
